Formatting

+++ tests/behaviour/contracts/cross_contract_calls/other_contract_same_type.sol
@@ -3,25 +3,35 @@ pragma solidity ^0.8.10;
 // SPDX-License-Identifier: MIT
 
 contract WARP {
-  uint public counter;
-	function getAndIncrementCounter() external returns (uint) { return counter++; }
-	function getAndIncrementOtherCounter(address addr) external returns (uint) {
-		WARP c = WARP(addr);
-		return c.getAndIncrementCounter();
-	}
-  function getCounters(address addr) external view returns (uint, uint) {
+  uint256 public counter;
+
+  function getAndIncrementCounter() external returns (uint256) {
+    return counter++;
+  }
+
+  function getAndIncrementOtherCounter(address addr) external returns (uint256) {
+    WARP c = WARP(addr);
+    return c.getAndIncrementCounter();
+  }
+
+  function getCounters(address addr) external view returns (uint256, uint256) {
     return (counter, WARP(addr).counter());
   }
 }
 
 contract WARPDuplicate {
-  uint public counter;
-	function getAndIncrementCounter() external returns (uint) { return counter++; }
-	function getAndIncrementOtherCounter(address addr) external returns (uint) {
-		WARP c = WARP(addr);
-		return c.getAndIncrementCounter();
-	}
-  function getCounters(address addr) external view returns (uint, uint) {
+  uint256 public counter;
+
+  function getAndIncrementCounter() external returns (uint256) {
+    return counter++;
+  }
+
+  function getAndIncrementOtherCounter(address addr) external returns (uint256) {
+    WARP c = WARP(addr);
+    return c.getAndIncrementCounter();
+  }
+
+  function getCounters(address addr) external view returns (uint256, uint256) {
     return (counter, WARP(addr).counter());
   }
 }
